apps/worker/requirements.txt
python-dotenv
google-generativeai
assemblyai
redis
numpy
Flask
psycopg2-binary
